Extract run-summary tests and tighten test CI
