+++ tests/test_cli.py
@@ -6,8 +6,7 @@ import os
 import subprocess
 import sys
 import tempfile
-import textwrap
-from unittest.mock import MagicMock, Mock, patch
+from unittest.mock import MagicMock, patch
 
 import pytest
 
@@ -3730,1304 +3729,6 @@ class TestDatasetsPermissionsFallback:
         assert output["dataViews"][0]["connection"]["name"] is None
         assert output["dataViews"][0]["connection"]["id"] == "dg_abc"
         assert output["dataViews"][0]["datasets"] == []
-
-
-class TestRunSummaryOutput:
-    """Tests for --run-summary-json output."""
-
-    @staticmethod
-    def _assert_run_summary_schema(payload):
-        """Validate run summary payload contract used by automation."""
-        required_keys = {
-            "summary_version",
-            "tool_version",
-            "started_at",
-            "ended_at",
-            "duration_seconds",
-            "exit_code",
-            "status",
-            "mode",
-            "profile",
-            "config_file",
-            "output_format",
-            "allow_partial",
-            "command",
-            "inputs",
-            "results",
-            "result_counts",
-            "failure_rollups",
-            "quality_gate_failed",
-            "quality_policy",
-            "details",
-        }
-        assert required_keys.issubset(payload)
-        assert payload["summary_version"] == "1.1"
-        assert isinstance(payload["tool_version"], str)
-        assert isinstance(payload["started_at"], str)
-        assert isinstance(payload["ended_at"], str)
-        assert isinstance(payload["duration_seconds"], (int, float))
-        assert isinstance(payload["exit_code"], int)
-        assert payload["status"] in {"success", "error", "policy_exit"}
-        assert isinstance(payload["mode"], str)
-        assert payload["profile"] is None or isinstance(payload["profile"], str)
-        assert payload["config_file"] is None or isinstance(payload["config_file"], str)
-        assert payload["output_format"] is None or isinstance(payload["output_format"], str)
-        assert isinstance(payload["allow_partial"], bool)
-
-        command = payload["command"]
-        assert isinstance(command, dict)
-        assert isinstance(command.get("argv"), list)
-        assert all(isinstance(arg, str) for arg in command["argv"])
-        assert isinstance(command.get("cwd"), str)
-
-        inputs = payload["inputs"]
-        assert isinstance(inputs, dict)
-        assert isinstance(inputs.get("data_view_inputs"), list)
-        assert isinstance(inputs.get("resolved_data_views"), list)
-
-        result_counts = payload["result_counts"]
-        assert isinstance(result_counts, dict)
-        assert isinstance(result_counts.get("total"), int)
-        assert isinstance(result_counts.get("successful"), int)
-        assert isinstance(result_counts.get("failed"), int)
-        assert isinstance(result_counts.get("quality_issues"), int)
-        assert result_counts["total"] == result_counts["successful"] + result_counts["failed"]
-
-        assert isinstance(payload["results"], list)
-        result_required_keys = {
-            "data_view_id",
-            "data_view_name",
-            "success",
-            "duration_seconds",
-            "metrics_count",
-            "dimensions_count",
-            "dq_issues_count",
-            "dq_severity_counts",
-            "output_file",
-            "output_files",
-            "error_message",
-            "failure_code",
-            "failure_reason",
-            "partial_output",
-            "partial_reasons",
-            "file_size_bytes",
-            "segments_count",
-            "segments_high_complexity",
-            "calculated_metrics_count",
-            "calculated_metrics_high_complexity",
-            "derived_fields_count",
-            "derived_fields_high_complexity",
-        }
-        for result in payload["results"]:
-            assert isinstance(result, dict)
-            assert result_required_keys.issubset(result)
-            assert isinstance(result["data_view_id"], str)
-            assert isinstance(result["data_view_name"], str)
-            assert isinstance(result["success"], bool)
-            assert isinstance(result["duration_seconds"], (int, float))
-            assert isinstance(result["dq_severity_counts"], dict)
-            assert isinstance(result["output_file"], str)
-            assert isinstance(result["output_files"], list)
-            assert all(isinstance(path, str) for path in result["output_files"])
-            assert isinstance(result["failure_code"], str)
-            assert isinstance(result["failure_reason"], str)
-            assert isinstance(result["partial_output"], bool)
-            assert isinstance(result["partial_reasons"], list)
-            assert all(isinstance(reason, str) for reason in result["partial_reasons"])
-
-        failure_rollups = payload["failure_rollups"]
-        assert isinstance(failure_rollups, dict)
-        assert isinstance(failure_rollups.get("by_code"), dict)
-        assert isinstance(failure_rollups.get("by_reason"), dict)
-
-        assert isinstance(payload["quality_gate_failed"], bool)
-
-        quality_policy = payload["quality_policy"]
-        if quality_policy is not None:
-            assert isinstance(quality_policy, dict)
-            assert isinstance(quality_policy.get("path"), str)
-            assert isinstance(quality_policy.get("applied"), dict)
-
-        assert isinstance(payload["details"], dict)
-
-    @patch("cja_auto_sdr.generator.process_single_dataview")
-    @patch("cja_auto_sdr.generator.resolve_data_view_names")
-    def test_run_summary_written_for_sdr_success(self, mock_resolve, mock_process, tmp_path):
-        """Successful SDR run should write run summary with result details."""
-        from cja_auto_sdr.generator import ProcessingResult, main
-
-        mock_resolve.return_value = (["dv_test"], {})
-        mock_process.return_value = ProcessingResult(
-            data_view_id="dv_test",
-            data_view_name="Test View",
-            success=True,
-            duration=0.25,
-            metrics_count=10,
-            dimensions_count=12,
-            dq_issues_count=0,
-            dq_issues=[],
-            dq_severity_counts={},
-            output_file="report.xlsx",
-            file_size_bytes=2048,
-        )
-
-        summary_file = tmp_path / "run_summary.json"
-        with patch.object(sys, "argv", ["cja_auto_sdr", "dv_test", "--run-summary-json", str(summary_file)]):
-            main()
-
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["mode"] == "sdr"
-        assert payload["exit_code"] == 0
-        assert payload["output_format"] == "excel"
-        assert payload["result_counts"]["total"] == 1
-        assert payload["result_counts"]["successful"] == 1
-        assert payload["results"][0]["data_view_id"] == "dv_test"
-        assert payload["results"][0]["output_file"] == "report.xlsx"
-        assert payload["results"][0]["output_files"] == ["report.xlsx"]
-
-    @patch("cja_auto_sdr.generator.process_single_dataview")
-    @patch("cja_auto_sdr.generator.resolve_data_view_names")
-    def test_run_summary_serializes_additive_output_files(self, mock_resolve, mock_process, tmp_path):
-        """Run summary should carry additive multi-artifact output data."""
-        from cja_auto_sdr.generator import ProcessingResult, main
-
-        mock_resolve.return_value = (["dv_test"], {})
-        mock_process.return_value = ProcessingResult(
-            data_view_id="dv_test",
-            data_view_name="Test View",
-            success=True,
-            duration=0.25,
-            metrics_count=10,
-            dimensions_count=12,
-            dq_issues_count=0,
-            dq_issues=[],
-            dq_severity_counts={},
-            output_file="report.xlsx",
-            output_files=["report.xlsx", "report.json", "report.html"],
-            file_size_bytes=2048,
-        )
-
-        summary_file = tmp_path / "run_summary_outputs.json"
-        with patch.object(sys, "argv", ["cja_auto_sdr", "dv_test", "--run-summary-json", str(summary_file)]):
-            main()
-
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["results"][0]["output_file"] == "report.xlsx"
-        assert payload["results"][0]["output_files"] == ["report.xlsx", "report.json", "report.html"]
-
-    def test_run_summary_all_format_subprocess_preserves_output_files(self, tmp_path):
-        """E2E: subprocess run-summary contract should preserve additive output_files for --format all."""
-        repo_root = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-        summary_file = tmp_path / "run_summary_all_subprocess.json"
-        script = textwrap.dedent(
-            f"""
-            import sys
-            from unittest.mock import patch
-
-            from cja_auto_sdr.generator import ProcessingResult, main
-
-            summary_file = {str(summary_file)!r}
-            result = ProcessingResult(
-                data_view_id="dv_test",
-                data_view_name="Test View",
-                success=True,
-                duration=0.25,
-                metrics_count=10,
-                dimensions_count=12,
-                dq_issues_count=0,
-                dq_issues=[],
-                dq_severity_counts={{}},
-                output_file="report.xlsx",
-                output_files=["report.xlsx", "report.json", "report.html"],
-                file_size_bytes=2048,
-            )
-
-            with (
-                patch("cja_auto_sdr.generator.resolve_data_view_names", return_value=(["dv_test"], {{}})),
-                patch("cja_auto_sdr.generator.process_single_dataview", return_value=result),
-                patch.object(
-                    sys,
-                    "argv",
-                    ["cja_auto_sdr", "dv_test", "--format", "all", "--run-summary-json", summary_file],
-                ),
-            ):
-                main()
-            """,
-        )
-
-        result = subprocess.run(
-            ["uv", "run", "python", "-c", script],
-            capture_output=True,
-            text=True,
-            cwd=repo_root,
-        )
-
-        assert result.returncode == 0, result.stderr
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["output_format"] == "all"
-        assert payload["results"][0]["output_file"] == "report.xlsx"
-        assert payload["results"][0]["output_files"] == ["report.xlsx", "report.json", "report.html"]
-
-
-class TestOpenOutputArtifacts(TestRunSummaryOutput):
-    """Tests for opening normalized emitted artifact paths."""
-
-    @patch("cja_auto_sdr.generator.open_file_in_default_app")
-    @patch("cja_auto_sdr.generator.process_single_dataview")
-    @patch("cja_auto_sdr.generator.resolve_data_view_names")
-    def test_single_mode_open_uses_all_emitted_output_files(self, mock_resolve, mock_process, mock_open):
-        """Single-mode --open should consume additive output_files when present."""
-        from cja_auto_sdr.generator import ProcessingResult, main
-
-        mock_resolve.return_value = (["dv_test"], {})
-        mock_process.return_value = ProcessingResult(
-            data_view_id="dv_test",
-            data_view_name="Test View",
-            success=True,
-            duration=0.25,
-            metrics_count=10,
-            dimensions_count=12,
-            dq_issues_count=0,
-            dq_issues=[],
-            dq_severity_counts={},
-            output_file="report.xlsx",
-            output_files=["report.xlsx", "report.json", "report.html"],
-            file_size_bytes=2048,
-        )
-        mock_open.return_value = True
-
-        with patch.object(sys, "argv", ["cja_auto_sdr", "dv_test", "--open", "--format", "all"]):
-            main()
-
-        assert mock_open.call_count == 3
-        assert [call.args[0] for call in mock_open.call_args_list] == ["report.xlsx", "report.json", "report.html"]
-
-    @patch("cja_auto_sdr.generator.process_single_dataview")
-    @patch("cja_auto_sdr.generator.resolve_data_view_names")
-    def test_run_summary_written_for_policy_exit(self, mock_resolve, mock_process, tmp_path):
-        """Policy exits should still write run summary with quality gate status."""
-        from cja_auto_sdr.generator import ProcessingResult, main
-
-        mock_resolve.return_value = (["dv_test"], {})
-        mock_process.return_value = ProcessingResult(
-            data_view_id="dv_test",
-            data_view_name="Test View",
-            success=True,
-            duration=0.1,
-            dq_issues_count=1,
-            dq_issues=[{"Severity": "HIGH", "Issue": "Duplicate component"}],
-            dq_severity_counts={"HIGH": 1},
-        )
-
-        summary_file = tmp_path / "run_summary_policy.json"
-        with patch.object(
-            sys,
-            "argv",
-            ["cja_auto_sdr", "dv_test", "--fail-on-quality", "HIGH", "--run-summary-json", str(summary_file)],
-        ):
-            with pytest.raises(SystemExit) as exc_info:
-                main()
-
-        assert exc_info.value.code == 2
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["exit_code"] == 2
-        assert payload["status"] == "policy_exit"
-        assert payload["quality_gate_failed"] is True
-
-    @patch("cja_auto_sdr.generator.process_single_dataview")
-    @patch("cja_auto_sdr.generator.resolve_data_view_names")
-    def test_run_summary_includes_stable_failure_identity(self, mock_resolve, mock_process, tmp_path):
-        """Failed SDR results should expose stable failure code/reason and rollups."""
-        from cja_auto_sdr.generator import ProcessingResult, main
-
-        mock_resolve.return_value = (["dv_test"], {})
-        mock_process.return_value = ProcessingResult(
-            data_view_id="dv_test",
-            data_view_name="Test View",
-            success=False,
-            duration=0.2,
-            error_message="Component fetch failed: metrics: timeout",
-            failure_code="COMPONENT_FETCH_FAILED",
-            failure_reason="required_endpoints_failed:metrics",
-        )
-
-        summary_file = tmp_path / "run_summary_failure_identity.json"
-        with patch.object(sys, "argv", ["cja_auto_sdr", "dv_test", "--run-summary-json", str(summary_file)]):
-            with pytest.raises(SystemExit) as exc_info:
-                main()
-
-        assert exc_info.value.code == 1
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["status"] == "error"
-        assert payload["result_counts"]["failed"] == 1
-        assert payload["results"][0]["failure_code"] == "COMPONENT_FETCH_FAILED"
-        assert payload["results"][0]["failure_reason"] == "required_endpoints_failed:metrics"
-        assert payload["failure_rollups"]["by_code"] == {"COMPONENT_FETCH_FAILED": 1}
-        assert payload["failure_rollups"]["by_reason"] == {"required_endpoints_failed:metrics": 1}
-
-    @patch("cja_auto_sdr.generator.process_single_dataview")
-    @patch("cja_auto_sdr.generator.resolve_data_view_names")
-    def test_run_summary_includes_partial_output_fields(self, mock_resolve, mock_process, tmp_path):
-        """Allow-partial SDR outputs should surface partial_output and partial_reasons per result."""
-        from cja_auto_sdr.generator import ProcessingResult, main
-
-        mock_resolve.return_value = (["dv_test"], {})
-        mock_process.return_value = ProcessingResult(
-            data_view_id="dv_test",
-            data_view_name="Test View",
-            success=True,
-            duration=0.2,
-            partial_output=True,
-            partial_reasons=[
-                "required_endpoints_failed:metrics",
-                "data_quality_validation_runtime_failed",
-            ],
-        )
-
-        summary_file = tmp_path / "run_summary_partial_output.json"
-        with patch.object(
-            sys,
-            "argv",
-            [
-                "cja_auto_sdr",
-                "dv_test",
-                "--allow-partial",
-                "--run-summary-json",
-                str(summary_file),
-            ],
-        ):
-            main()
-
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["allow_partial"] is True
-        assert payload["results"][0]["partial_output"] is True
-        assert payload["results"][0]["partial_reasons"] == [
-            "required_endpoints_failed:metrics",
-            "data_quality_validation_runtime_failed",
-        ]
-
-    @patch("cja_auto_sdr.generator.resolve_data_view_names")
-    @patch("cja_auto_sdr.generator.BatchProcessor.print_summary")
-    @patch("cja_auto_sdr.generator.ProcessPoolExecutor")
-    @patch("cja_auto_sdr.generator.tqdm")
-    def test_run_summary_batch_failure_rollups_mixed_results_under_concurrent_completion(
-        self,
-        mock_tqdm,
-        mock_executor_cls,
-        _mock_print_summary,
-        mock_resolve,
-        tmp_path,
-    ):
-        """Batch-mode run summary should aggregate failure_rollups with mixed out-of-order future completion."""
-        from cja_auto_sdr.generator import ProcessingResult, main
-
-        mock_resolve.return_value = (["dv_ok", "dv_fetch_fail", "dv_validation_fail"], {})
-
-        mock_pbar = MagicMock()
-        mock_tqdm.return_value.__enter__ = Mock(return_value=mock_pbar)
-        mock_tqdm.return_value.__exit__ = Mock(return_value=False)
-
-        future_ok = Mock()
-        future_fetch_fail = Mock()
-        future_validation_fail = Mock()
-
-        future_ok.result.return_value = ProcessingResult(
-            data_view_id="dv_ok",
-            data_view_name="Healthy View",
-            success=True,
-            duration=0.2,
-            dq_issues_count=1,
-            dq_issues=[{"Severity": "HIGH", "Issue": "threshold"}],
-            dq_severity_counts={"HIGH": 1},
-        )
-        future_fetch_fail.result.return_value = ProcessingResult(
-            data_view_id="dv_fetch_fail",
-            data_view_name="Fetch Fail View",
-            success=False,
-            duration=0.3,
-            error_message="Component fetch failed: metrics: timeout",
-            failure_code="COMPONENT_FETCH_FAILED",
-            failure_reason="required_endpoints_failed:metrics",
-        )
-        future_validation_fail.result.return_value = ProcessingResult(
-            data_view_id="dv_validation_fail",
-            data_view_name="Validation Fail View",
-            success=False,
-            duration=0.3,
-            error_message="Data quality validation failed: threadpool failure",
-            failure_code="DQ_VALIDATION_RUNTIME_FAILED",
-            failure_reason="data_quality_validation_runtime_failed",
-        )
-
-        mock_executor = MagicMock()
-        mock_executor.__enter__ = Mock(return_value=mock_executor)
-        mock_executor.__exit__ = Mock(return_value=False)
-        mock_executor.submit.side_effect = [future_ok, future_fetch_fail, future_validation_fail]
-        mock_executor_cls.return_value = mock_executor
-
-        summary_file = tmp_path / "run_summary_batch_rollups.json"
-        with (
-            patch(
-                "cja_auto_sdr.generator.as_completed",
-                return_value=[future_validation_fail, future_ok, future_fetch_fail],
-            ),
-            patch("cja_auto_sdr.generator.setup_logging", return_value=Mock()),
-            patch.object(
-                sys,
-                "argv",
-                [
-                    "cja_auto_sdr",
-                    "dv_ok",
-                    "dv_fetch_fail",
-                    "dv_validation_fail",
-                    "--batch",
-                    "--workers",
-                    "2",
-                    "--continue-on-error",
-                    "--fail-on-quality",
-                    "HIGH",
-                    "--run-summary-json",
-                    str(summary_file),
-                ],
-            ),
-        ):
-            with pytest.raises(SystemExit) as exc_info:
-                main()
-
-        assert exc_info.value.code == 1
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["result_counts"]["failed"] == 2
-        assert payload["failure_rollups"]["by_code"] == {
-            "COMPONENT_FETCH_FAILED": 1,
-            "DQ_VALIDATION_RUNTIME_FAILED": 1,
-        }
-        assert payload["failure_rollups"]["by_reason"] == {
-            "data_quality_validation_runtime_failed": 1,
-            "required_endpoints_failed:metrics": 1,
-        }
-
-    @patch("cja_auto_sdr.generator.resolve_data_view_names")
-    @patch("cja_auto_sdr.generator.BatchProcessor.print_summary")
-    @patch("cja_auto_sdr.generator.ProcessPoolExecutor")
-    @patch("cja_auto_sdr.generator.tqdm")
-    def test_run_summary_batch_mixed_results_preserves_success_output_files(
-        self,
-        mock_tqdm,
-        mock_executor_cls,
-        _mock_print_summary,
-        mock_resolve,
-        tmp_path,
-    ):
-        """Batch-mode run summary should preserve additive output_files for successful entries even when failures exist."""
-        from cja_auto_sdr.generator import ProcessingResult, main
-
-        mock_resolve.return_value = (["dv_ok", "dv_fail"], {})
-
-        mock_pbar = MagicMock()
-        mock_tqdm.return_value.__enter__ = Mock(return_value=mock_pbar)
-        mock_tqdm.return_value.__exit__ = Mock(return_value=False)
-
-        future_ok = Mock()
-        future_fail = Mock()
-
-        future_ok.result.return_value = ProcessingResult(
-            data_view_id="dv_ok",
-            data_view_name="Healthy View",
-            success=True,
-            duration=0.2,
-            metrics_count=10,
-            dimensions_count=12,
-            dq_issues_count=0,
-            dq_issues=[],
-            dq_severity_counts={},
-            output_file="ok.xlsx",
-            output_files=["ok.json", "ok.xlsx", "ok.html", "ok.json"],
-            file_size_bytes=2048,
-        )
-        future_fail.result.return_value = ProcessingResult(
-            data_view_id="dv_fail",
-            data_view_name="Fail View",
-            success=False,
-            duration=0.3,
-            error_message="Component fetch failed: metrics: timeout",
-            failure_code="COMPONENT_FETCH_FAILED",
-            failure_reason="required_endpoints_failed:metrics",
-        )
-
-        mock_executor = MagicMock()
-        mock_executor.__enter__ = Mock(return_value=mock_executor)
-        mock_executor.__exit__ = Mock(return_value=False)
-        mock_executor.submit.side_effect = [future_ok, future_fail]
-        mock_executor_cls.return_value = mock_executor
-
-        summary_file = tmp_path / "run_summary_batch_outputs.json"
-        with (
-            patch("cja_auto_sdr.generator.as_completed", return_value=[future_fail, future_ok]),
-            patch("cja_auto_sdr.generator.setup_logging", return_value=Mock()),
-            patch.object(
-                sys,
-                "argv",
-                [
-                    "cja_auto_sdr",
-                    "dv_ok",
-                    "dv_fail",
-                    "--batch",
-                    "--workers",
-                    "2",
-                    "--continue-on-error",
-                    "--run-summary-json",
-                    str(summary_file),
-                ],
-            ),
-        ):
-            main()
-
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        results_by_id = {result["data_view_id"]: result for result in payload["results"]}
-        assert results_by_id["dv_ok"]["output_file"] == "ok.xlsx"
-        assert results_by_id["dv_ok"]["output_files"] == ["ok.xlsx", "ok.json", "ok.html"]
-        assert results_by_id["dv_fail"]["output_files"] == []
-
-    @patch("cja_auto_sdr.generator.list_dataviews")
-    def test_run_summary_written_for_discovery_mode(self, mock_list_dataviews, tmp_path):
-        """Discovery mode should write summary even when exiting via SystemExit."""
-        from cja_auto_sdr.generator import main
-
-        mock_list_dataviews.return_value = True
-        summary_file = tmp_path / "run_summary_discovery.json"
-
-        with patch.object(
-            sys,
-            "argv",
-            ["cja_auto_sdr", "--list-dataviews", "--run-summary-json", str(summary_file)],
-        ):
-            with pytest.raises(SystemExit) as exc_info:
-                main()
-
-        assert exc_info.value.code == 0
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["mode"] == "discovery"
-        assert payload["exit_code"] == 0
-        assert payload["details"]["discovery_command"] == "list_dataviews"
-
-    @patch("cja_auto_sdr.generator.list_dataviews")
-    def test_run_summary_non_sdr_quality_policy_is_not_applied(self, mock_list_dataviews, tmp_path):
-        """Run summary should record quality policy path but keep applied defaults empty in non-SDR modes."""
-        from cja_auto_sdr.generator import main
-
-        policy_path = tmp_path / "quality_policy.json"
-        policy_path.write_text(
-            json.dumps({"fail_on_quality": "HIGH", "quality_report": "csv", "max_issues": 5}),
-            encoding="utf-8",
-        )
-        mock_list_dataviews.return_value = True
-        summary_file = tmp_path / "run_summary_discovery_quality_policy.json"
-
-        with patch.object(
-            sys,
-            "argv",
-            [
-                "cja_auto_sdr",
-                "--list-dataviews",
-                "--quality-policy",
-                str(policy_path),
-                "--run-summary-json",
-                str(summary_file),
-            ],
-        ):
-            with pytest.raises(SystemExit) as exc_info:
-                main()
-
-        assert exc_info.value.code == 0
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["mode"] == "discovery"
-        assert payload["quality_policy"]["path"] == str(policy_path)
-        assert payload["quality_policy"]["applied"] == {}
-
-    @patch("cja_auto_sdr.generator.run_org_report")
-    @patch("cja_auto_sdr.generator.list_dataviews")
-    def test_run_summary_mode_precedence_matches_dispatch_order(
-        self,
-        mock_list_dataviews,
-        mock_run_org_report,
-        tmp_path,
-    ):
-        """When multiple mode flags are present, summary mode should match the first dispatch branch."""
-        from cja_auto_sdr.generator import main
-
-        mock_list_dataviews.return_value = True
-        summary_file = tmp_path / "run_summary_mode_precedence.json"
-
-        with patch.object(
-            sys,
-            "argv",
-            ["cja_auto_sdr", "--list-dataviews", "--org-report", "--run-summary-json", str(summary_file)],
-        ):
-            with pytest.raises(SystemExit) as exc_info:
-                main()
-
-        assert exc_info.value.code == 0
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["mode"] == "discovery"
-        assert payload["details"]["discovery_command"] == "list_dataviews"
-        mock_list_dataviews.assert_called_once()
-        mock_run_org_report.assert_not_called()
-
-    @patch("cja_auto_sdr.generator.process_single_dataview")
-    @patch("cja_auto_sdr.generator.resolve_data_view_names")
-    def test_run_summary_stdout_is_json_only(self, mock_resolve, mock_process, capsys):
-        """--run-summary-json stdout should emit parseable JSON on stdout."""
-        from cja_auto_sdr.generator import ProcessingResult, main
-
-        mock_resolve.return_value = (["dv_test"], {})
-        mock_process.return_value = ProcessingResult(
-            data_view_id="dv_test",
-            data_view_name="Test View",
-            success=True,
-            duration=0.1,
-            metrics_count=1,
-            dimensions_count=1,
-            dq_issues_count=0,
-            dq_issues=[],
-            dq_severity_counts={},
-        )
-
-        with patch.object(sys, "argv", ["cja_auto_sdr", "dv_test", "--run-summary-json", "stdout"]):
-            main()
-
-        captured = capsys.readouterr()
-        payload = json.loads(captured.out)
-        self._assert_run_summary_schema(payload)
-        assert payload["mode"] == "sdr"
-        assert payload["exit_code"] == 0
-
-    @patch("cja_auto_sdr.generator.process_single_dataview")
-    @patch("cja_auto_sdr.generator.resolve_data_view_names")
-    @patch("cja_auto_sdr.generator.interactive_wizard")
-    def test_run_summary_interactive_refreshes_data_view_inputs(
-        self,
-        mock_wizard,
-        mock_resolve,
-        mock_process,
-        tmp_path,
-    ):
-        """Interactive runs should record wizard-selected data view inputs in run summary."""
-        from cja_auto_sdr.generator import ProcessingResult, WizardConfig, main
-
-        mock_wizard.return_value = WizardConfig(
-            data_view_ids=["dv_wizard_selected"],
-            output_format="excel",
-            include_segments=False,
-            include_calculated=False,
-            include_derived=False,
-            inventory_only=False,
-        )
-        mock_resolve.return_value = (["dv_wizard_selected"], {})
-        mock_process.return_value = ProcessingResult(
-            data_view_id="dv_wizard_selected",
-            data_view_name="Wizard Selected",
-            success=True,
-            duration=0.1,
-            metrics_count=1,
-            dimensions_count=1,
-            dq_issues_count=0,
-            dq_issues=[],
-            dq_severity_counts={},
-        )
-
-        summary_file = tmp_path / "run_summary_interactive_inputs.json"
-        with patch.object(sys, "argv", ["cja_auto_sdr", "--interactive", "--run-summary-json", str(summary_file)]):
-            main()
-
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["mode"] == "sdr"
-        assert payload["inputs"]["data_view_inputs"] == ["dv_wizard_selected"]
-
-    @patch("cja_auto_sdr.generator.list_dataviews")
-    def test_run_summary_stdout_with_abbreviated_flag_is_json_only(self, mock_list_dataviews, capsys):
-        """Abbreviated --run-summary-json should still redirect normal stdout chatter."""
-        from cja_auto_sdr.generator import main
-
-        def _mock_list_dataviews(*args, **kwargs):
-            print("discovery table output")
-            return True
-
-        mock_list_dataviews.side_effect = _mock_list_dataviews
-
-        with patch.object(sys, "argv", ["cja_auto_sdr", "--list-dataviews", "--run-summary-j", "stdout"]):
-            with pytest.raises(SystemExit) as exc_info:
-                main()
-
-        assert exc_info.value.code == 0
-        captured = capsys.readouterr()
-        payload = json.loads(captured.out)
-        self._assert_run_summary_schema(payload)
-        assert payload["mode"] == "discovery"
-        assert payload["exit_code"] == 0
-        assert "discovery table output" not in captured.out
-        assert "discovery table output" in captured.err
-
-    def test_run_summary_stdout_json_only_subprocess_exit_codes_full_flag(self):
-        """E2E: full flag should keep stdout as JSON-only in chatty exit-codes mode."""
-        repo_root = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-        result = subprocess.run(
-            ["uv", "run", "cja_auto_sdr", "--exit-codes", "--run-summary-json", "stdout"],
-            capture_output=True,
-            text=True,
-            cwd=repo_root,
-        )
-        assert result.returncode == 0
-        payload = json.loads(result.stdout)
-        self._assert_run_summary_schema(payload)
-        assert payload["mode"] == "exit_codes"
-        assert "EXIT CODE REFERENCE" not in result.stdout
-        assert "EXIT CODE REFERENCE" in result.stderr
-
-    def test_run_summary_stdout_json_only_subprocess_exit_codes_abbreviated_flag(self):
-        """E2E: abbreviated flag should keep stdout as JSON-only in chatty exit-codes mode."""
-        repo_root = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-        result = subprocess.run(
-            ["uv", "run", "cja_auto_sdr", "--exit-codes", "--run-summary-j", "stdout"],
-            capture_output=True,
-            text=True,
-            cwd=repo_root,
-        )
-        assert result.returncode == 0
-        payload = json.loads(result.stdout)
-        self._assert_run_summary_schema(payload)
-        assert payload["mode"] == "exit_codes"
-        assert "EXIT CODE REFERENCE" not in result.stdout
-        assert "EXIT CODE REFERENCE" in result.stderr
-
-    def test_run_summary_exit_codes_ignored_allow_partial_reports_false(self, tmp_path):
-        """Ignored standalone flags should not survive into run-summary telemetry."""
-        from cja_auto_sdr.generator import main
-
-        summary_file = tmp_path / "run_summary_exit_codes_allow_partial_ignored.json"
-        with patch.object(
-            sys,
-            "argv",
-            [
-                "cja_auto_sdr",
-                "--exit-codes",
-                "--allow-partial",
-                "--run-summary-json",
-                str(summary_file),
-            ],
-        ):
-            with pytest.raises(SystemExit) as exc_info:
-                main()
-
-        assert exc_info.value.code == 0
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["mode"] == "exit_codes"
-        assert payload["status"] == "success"
-        assert payload["allow_partial"] is False
-
-    def test_run_summary_completion_mode_classification(self, tmp_path):
-        """Completion runs should emit run summary mode=completion."""
-        from cja_auto_sdr.generator import main
-
-        summary_file = tmp_path / "run_summary_completion.json"
-        fake_argcomplete = type(sys)("argcomplete")
-
-        with (
-            patch.object(
-                sys,
-                "argv",
-                ["cja_auto_sdr", "--completion", "bash", "--run-summary-json", str(summary_file)],
-            ),
-            patch.dict("sys.modules", {"argcomplete": fake_argcomplete}),
-        ):
-            with pytest.raises(SystemExit) as exc_info:
-                main()
-
-        assert exc_info.value.code == 0
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["mode"] == "completion"
-        assert payload["exit_code"] == 0
-        assert payload["status"] == "success"
-
-    def test_run_summary_exit_codes_precedes_completion_mode_classification(self, tmp_path):
-        """Mixed --exit-codes/--completion should preserve exit-codes summary mode."""
-        from cja_auto_sdr.generator import main
-
-        summary_file = tmp_path / "run_summary_exit_codes_precedes_completion.json"
-        with (
-            patch.object(
-                sys,
-                "argv",
-                [
-                    "cja_auto_sdr",
-                    "--exit-codes",
-                    "--completion",
-                    "bash",
-                    "--run-summary-json",
-                    str(summary_file),
-                ],
-            ),
-            patch("cja_auto_sdr.core.exit_codes.print_exit_codes") as mock_print_exit_codes,
-            patch("cja_auto_sdr.generator._handle_completion_prevalidation") as mock_completion_prevalidation,
-        ):
-            with pytest.raises(SystemExit) as exc_info:
-                main()
-
-        assert exc_info.value.code == 0
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["mode"] == "exit_codes"
-        assert payload["exit_code"] == 0
-        assert payload["status"] == "success"
-        mock_print_exit_codes.assert_called_once()
-        mock_completion_prevalidation.assert_not_called()
-
-    def test_run_summary_stdout_subprocess_version_is_order_independent(self):
-        """E2E: --version should still emit run summary JSON regardless of flag order."""
-        repo_root = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-        commands = [
-            ["uv", "run", "cja_auto_sdr", "--version", "--run-summary-json", "stdout"],
-            ["uv", "run", "cja_auto_sdr", "--run-summary-json", "stdout", "--version"],
-            ["uv", "run", "cja_auto_sdr", "-V", "--run-summary-json", "stdout"],
-            ["uv", "run", "cja_auto_sdr", "--run-summary-json", "stdout", "-V"],
-            ["uv", "run", "cja_auto_sdr", "--version", "--profile", "--run-summary-json", "stdout"],
-        ]
-
-        for cmd in commands:
-            result = subprocess.run(
-                cmd,
-                capture_output=True,
-                text=True,
-                cwd=repo_root,
-            )
-            assert result.returncode == 0
-            payload = json.loads(result.stdout)
-            self._assert_run_summary_schema(payload)
-            assert payload["exit_code"] == 0
-            assert payload["mode"] == "unknown"
-            assert "cja_auto_sdr " not in result.stdout
-            assert "cja_auto_sdr " in result.stderr
-
-    def test_module_invocation_version_banner_consistent_with_and_without_run_summary(self):
-        """`python -m cja_auto_sdr --version` should keep the same banner prefix across paths."""
-        repo_root = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-
-        fast_path = subprocess.run(
-            ["uv", "run", "python", "-m", "cja_auto_sdr", "--version"],
-            capture_output=True,
-            text=True,
-            cwd=repo_root,
-        )
-        assert fast_path.returncode == 0
-        assert " -m cja_auto_sdr " in fast_path.stdout
-        fast_prefix = fast_path.stdout.strip().split(" -m cja_auto_sdr ")[0]
-
-        fallback = subprocess.run(
-            ["uv", "run", "python", "-m", "cja_auto_sdr", "--version", "--run-summary-json", "stdout"],
-            capture_output=True,
-            text=True,
-            cwd=repo_root,
-        )
-        assert fallback.returncode == 0
-        payload = json.loads(fallback.stdout)
-        self._assert_run_summary_schema(payload)
-        assert payload["exit_code"] == 0
-        assert f"{fast_prefix} -m cja_auto_sdr " in fallback.stderr
-
-    @patch("cja_auto_sdr.generator.process_single_dataview")
-    @patch("cja_auto_sdr.generator.resolve_data_view_names")
-    def test_run_summary_records_inferred_output_format(self, mock_resolve, mock_process, tmp_path):
-        """Run summary should capture format inferred from --output extension."""
-        from cja_auto_sdr.generator import ProcessingResult, main
-
-        mock_resolve.return_value = (["dv_test"], {})
-        mock_process.return_value = ProcessingResult(
-            data_view_id="dv_test",
-            data_view_name="Test View",
-            success=True,
-            duration=0.1,
-            metrics_count=1,
-            dimensions_count=1,
-            dq_issues_count=0,
-            dq_issues=[],
-            dq_severity_counts={},
-            output_file="report.csv",
-        )
-
-        summary_file = tmp_path / "run_summary_inferred.json"
-        with patch.object(
-            sys,
-            "argv",
-            ["cja_auto_sdr", "dv_test", "--output", "report.csv", "--run-summary-json", str(summary_file)],
-        ):
-            main()
-
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["output_format"] == "csv"
-
-    @patch("cja_auto_sdr.generator.write_quality_report_output")
-    @patch("cja_auto_sdr.generator.process_single_dataview")
-    @patch("cja_auto_sdr.generator.resolve_data_view_names")
-    def test_run_summary_quality_report_uses_quality_report_format(
-        self,
-        mock_resolve,
-        mock_process,
-        mock_write_report,
-        tmp_path,
-    ):
-        """Run summary output_format should reflect --quality-report format, not internal SDR writer format."""
-        from cja_auto_sdr.generator import ProcessingResult, main
-
-        mock_resolve.return_value = (["dv_test"], {})
-        mock_write_report.return_value = "stdout"
-        mock_process.return_value = ProcessingResult(
-            data_view_id="dv_test",
-            data_view_name="Test View",
-            success=True,
-            duration=0.1,
-            dq_issues_count=1,
-            dq_issues=[{"Severity": "LOW", "Issue": "Minor"}],
-            dq_severity_counts={"LOW": 1},
-        )
-
-        summary_file = tmp_path / "run_summary_quality_report_format.json"
-        with patch.object(
-            sys,
-            "argv",
-            ["cja_auto_sdr", "dv_test", "--quality-report", "csv", "--run-summary-json", str(summary_file)],
-        ):
-            main()
-
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["mode"] == "sdr"
-        assert payload["output_format"] == "csv"
-
-    @patch("cja_auto_sdr.generator.process_inventory_summary")
-    @patch("cja_auto_sdr.generator.resolve_data_view_names")
-    def test_run_summary_inventory_summary_mode_and_output_format(self, mock_resolve, mock_inventory_summary, tmp_path):
-        """Inventory summary runs should report mode=inventory_summary with effective summary output format."""
-        from cja_auto_sdr.generator import main
-
-        mock_resolve.return_value = (["dv_test"], {})
-        mock_inventory_summary.return_value = {}
-        summary_file = tmp_path / "run_summary_inventory_summary.json"
-
-        with patch.object(
-            sys,
-            "argv",
-            [
-                "cja_auto_sdr",
-                "dv_test",
-                "--include-segments",
-                "--inventory-summary",
-                "--run-summary-json",
-                str(summary_file),
-            ],
-        ):
-            with pytest.raises(SystemExit) as exc_info:
-                main()
-
-        assert exc_info.value.code == 0
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["mode"] == "inventory_summary"
-        assert payload["output_format"] == "console"
-
-    @patch("cja_auto_sdr.generator.process_inventory_summary")
-    @patch("cja_auto_sdr.generator.resolve_data_view_names")
-    def test_inventory_summary_propagates_log_format(self, mock_resolve, mock_inventory_summary):
-        """Inventory summary mode should pass --log-format to process_inventory_summary."""
-        from cja_auto_sdr.generator import main
-
-        mock_resolve.return_value = (["dv_test"], {})
-        mock_inventory_summary.return_value = {}
-
-        with patch.object(
-            sys,
-            "argv",
-            [
-                "cja_auto_sdr",
-                "dv_test",
-                "--include-segments",
-                "--inventory-summary",
-                "--log-format",
-                "json",
-            ],
-        ):
-            with pytest.raises(SystemExit) as exc_info:
-                main()
-
-        assert exc_info.value.code == 0
-        assert mock_inventory_summary.call_count == 1
-        assert mock_inventory_summary.call_args.kwargs["log_format"] == "json"
-
-    @patch("cja_auto_sdr.generator.git_init_snapshot_repo")
-    def test_run_summary_git_init_mode(self, mock_git_init, tmp_path):
-        """Run summary should classify --git-init runs with git_init mode."""
-        from cja_auto_sdr.generator import main
-
-        mock_git_init.return_value = (True, "initialized")
-        summary_file = tmp_path / "run_summary_git_init.json"
-
-        with patch.object(
-            sys,
-            "argv",
-            [
-                "cja_auto_sdr",
-                "--git-init",
-                "--git-dir",
-                str(tmp_path / "repo"),
-                "--run-summary-json",
-                str(summary_file),
-            ],
-        ):
-            with pytest.raises(SystemExit) as exc_info:
-                main()
-
-        assert exc_info.value.code == 0
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["mode"] == "git_init"
-        assert payload["status"] == "success"
-
-    def test_run_summary_invalid_cli_status_is_error(self, tmp_path):
-        """Argparse usage errors (exit 2) should be reported as status=error, not policy_exit."""
-        from cja_auto_sdr.generator import main
-
-        summary_file = tmp_path / "run_summary_cli_error.json"
-        with patch.object(
-            sys,
-            "argv",
-            ["cja_auto_sdr", "--definitely-invalid-flag", "--run-summary-json", str(summary_file)],
-        ):
-            with pytest.raises(SystemExit) as exc_info:
-                main()
-
-        assert exc_info.value.code == 2
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["status"] == "error"
-        assert payload["exit_code"] == 2
-
-    def test_run_summary_invalid_quality_policy_preserves_inferred_mode(self, tmp_path):
-        """Policy-load failures should still emit run summary with inferred mode metadata."""
-        from cja_auto_sdr.generator import main
-
-        summary_file = tmp_path / "run_summary_quality_policy_error.json"
-        missing_policy = tmp_path / "missing_quality_policy.json"
-
-        with patch.object(
-            sys,
-            "argv",
-            [
-                "cja_auto_sdr",
-                "--list-dataviews",
-                "--quality-policy",
-                str(missing_policy),
-                "--run-summary-json",
-                str(summary_file),
-            ],
-        ):
-            with pytest.raises(SystemExit) as exc_info:
-                main()
-
-        assert exc_info.value.code == 1
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["mode"] == "discovery"
-        assert payload["status"] == "error"
-        assert payload["quality_policy"]["path"] == str(missing_policy)
-        assert payload["quality_policy"]["applied"] == {}
-
-    def test_run_summary_policy_applied_allow_partial_survives_early_validation_exit(self, tmp_path):
-        """Policy-mutated allow_partial should be synced before later CLI validation exits."""
-        from cja_auto_sdr.generator import main
-
-        summary_file = tmp_path / "run_summary_policy_allow_partial_validation_error.json"
-        policy_file = tmp_path / "quality_policy.json"
-        policy_file.write_text(json.dumps({"allow_partial": True}), encoding="utf-8")
-
-        with patch.object(
-            sys,
-            "argv",
-            [
-                "cja_auto_sdr",
-                "dv_test",
-                "--quality-policy",
-                str(policy_file),
-                "--workers",
-                "0",
-                "--run-summary-json",
-                str(summary_file),
-            ],
-        ):
-            with pytest.raises(SystemExit) as exc_info:
-                main()
-
-        assert exc_info.value.code == 1
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["mode"] == "sdr"
-        assert payload["status"] == "error"
-        assert payload["allow_partial"] is True
-        assert payload["quality_policy"]["path"] == str(policy_file)
-        assert payload["quality_policy"]["applied"] == {"allow_partial": True}
-
-    def test_run_summary_profile_overwrite_validation_error_mode(self, tmp_path):
-        """Profile overwrite validation failures should still be classified as profile_management mode."""
-        from cja_auto_sdr.generator import main
-
-        summary_file = tmp_path / "run_summary_profile_overwrite_error.json"
-        with patch.object(
-            sys,
-            "argv",
-            ["cja_auto_sdr", "--profile-overwrite", "--run-summary-json", str(summary_file)],
-        ):
-            with pytest.raises(SystemExit) as exc_info:
-                main()
-
-        assert exc_info.value.code == 1
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["mode"] == "profile_management"
-        assert payload["status"] == "error"
-
-    def test_run_summary_non_sdr_allow_partial_validation_preserves_flag(self, tmp_path):
-        """Early non-SDR validation errors should preserve allow_partial telemetry."""
-        from cja_auto_sdr.generator import main
-
-        summary_file = tmp_path / "run_summary_non_sdr_allow_partial_error.json"
-        with patch.object(
-            sys,
-            "argv",
-            [
-                "cja_auto_sdr",
-                "--list-dataviews",
-                "--allow-partial",
-                "--run-summary-json",
-                str(summary_file),
-            ],
-        ):
-            with pytest.raises(SystemExit) as exc_info:
-                main()
-
-        assert exc_info.value.code == 1
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["mode"] == "discovery"
-        assert payload["status"] == "error"
-        assert payload["allow_partial"] is True
-
-    def test_run_summary_argparse_error_preserves_allow_partial_flag(self, tmp_path):
-        """Argparse failures should still preserve allow_partial telemetry from argv."""
-        from cja_auto_sdr.generator import main
-
-        summary_file = tmp_path / "run_summary_argparse_allow_partial_error.json"
-        with patch.object(
-            sys,
-            "argv",
-            [
-                "cja_auto_sdr",
-                "--allow-partial",
-                "--definitely-invalid-flag",
-                "--run-summary-json",
-                str(summary_file),
-            ],
-        ):
-            with pytest.raises(SystemExit) as exc_info:
-                main()
-
-        assert exc_info.value.code == 2
-        payload = json.loads(summary_file.read_text())
-        self._assert_run_summary_schema(payload)
-        assert payload["mode"] == "unknown"
-        assert payload["status"] == "error"
-        assert payload["allow_partial"] is True
-
-    def test_run_summary_missing_value_does_not_write_flag_named_file(self, tmp_path, monkeypatch):
-        """Malformed --run-summary-json should not treat the next flag as an output path."""
-        from cja_auto_sdr.generator import main
-
-        bad_output_name = "--list-dataviews"
-        monkeypatch.chdir(tmp_path)
-        with patch.object(sys, "argv", ["cja_auto_sdr", "--run-summary-json", "--list-dataviews"]):
-            with pytest.raises(SystemExit) as exc_info:
-                main()
-
-        assert exc_info.value.code == 2
-        assert not (tmp_path / bad_output_name).exists()
-
-    def test_cli_option_value_uses_last_valid_occurrence(self):
-        """Raw option helper should align with argparse 'last value wins' behavior."""
-        from cja_auto_sdr.generator import _cli_option_value
-
-        value = _cli_option_value(
-            "--run-summary-json",
-            [
-                "--run-summary-json",
-                "stdout",
-                "--run-summary-json",
-                "summary.json",
-            ],
-        )
-        assert value == "summary.json"
-
-    def test_cli_option_value_accepts_unambiguous_long_option_abbreviation(self):
-        """Raw option helper should resolve argparse-accepted long-option abbreviations."""
-        from cja_auto_sdr.generator import _cli_option_value
-
-        value = _cli_option_value("--run-summary-json", ["--run-summary-j", "stdout"])
-        assert value == "stdout"
-
-    @pytest.mark.parametrize(
-        ("argv", "expected"),
-        [
-            (["--run-summary-json", "stdout"], "stdout"),
-            (["--run-summary-j", "stdout"], "stdout"),
-            (["--run-summary-j=stdout"], "stdout"),
-            (["--run-summary-json", "--list-dataviews"], None),
-            (["--run-summary-j", "--list-dataviews"], None),
-            (["--run-summary-json", "stdout", "--run-summary-j", "summary.json"], "summary.json"),
-            (["--run-summary-json=stdout", "--run-summary-j=summary.json"], "summary.json"),
-            (["--run-summary-json", "stdout", "--run-summary-json"], "stdout"),
-        ],
-    )
-    def test_cli_option_value_permutations(self, argv, expected):
-        """Raw option helper should stay aligned with argparse-style option permutations."""
-        from cja_auto_sdr.generator import _cli_option_value
-
-        assert _cli_option_value("--run-summary-json", argv) == expected
-
-    @pytest.mark.parametrize(
-        ("option_name", "argv", "expected"),
-        [
-            ("--run-summary-json", ["--run-summary-j", "stdout"], True),
-            ("--max-issues", ["--max-i=10"], True),
-            ("--fail-on-quality", ["--fail-on-q", "HIGH"], True),
-            ("--profile", ["--pro"], False),
-            ("--run-summary-json", ["-q"], False),
-        ],
-    )
-    def test_cli_option_specified_permutations(self, option_name, argv, expected):
-        """Explicit option detection should follow argparse abbreviation semantics."""
-        from cja_auto_sdr.generator import _cli_option_specified
-
-        assert _cli_option_specified(option_name, argv) is expected
 
 
 class TestRunModeInference:
